Relocate mock testing utilities alongside other testing utilities
