+++ src/__tests__/fetchMore.ts
@@ -1,7 +1,7 @@
 import { assign, cloneDeep } from 'lodash';
 import gql from 'graphql-tag';
 
-import { mockSingleLink } from '../__mocks__/mockLinks';
+import { mockSingleLink } from './utils/mocks/mockLinks';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
 import { ApolloClient, NetworkStatus, ObservableQuery } from '../';
 import { itAsync } from './utils/itAsync';
