+++ src/__tests__/local-state/resolvers.ts
@@ -5,7 +5,7 @@ import { assign } from 'lodash';
 import { Observable, Observer } from '../../utilities/observables/Observable';
 import { ApolloLink } from '../../link/core/ApolloLink';
 import { ApolloClient } from '../..';
-import mockQueryManager from '../../__mocks__/mockQueryManager';
+import mockQueryManager from '../utils/mocks/mockQueryManager';
 import wrap from '../../__tests__/utils/wrap';
 import { itAsync } from '../utils/itAsync';
 import { ApolloQueryResult, Resolvers } from '../../core/types';
