+++ src/core/__tests__/QueryManager/live.ts
@@ -8,7 +8,7 @@
 import gql from 'graphql-tag';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
 import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
-import { MockSubscriptionLink } from '../../../__mocks__/mockLinks';
+import { MockSubscriptionLink } from '../../../__tests__/utils/mocks/mockLinks';
 
 // core
 import { QueryManager } from '../../QueryManager';
